# openapi.yaml
openapi: "3.1.2"
info:
  title: Komodo Accounts API
  version: "1.0.0"
  description: |
    Account profile, address, payment method, and preference management.

    **Two servers:**
    - Public (7051): account-facing profile endpoints; require user JWT
    - Internal (7052): service-to-service account lookup endpoints; require service JWT

    **Auth:** All `/v1/me/profile`, `/v1/me/addresses`, `/v1/me/payments`, and `/v1/me/preferences`
    routes require a user-scoped JWT (`Authorization: Bearer <token>`).
    Internal `/v1/accounts/{id}/*` routes require a service-scoped JWT (`svc:` prefix claim).

    **HTTP conventions:**
    - POST → 201 Created / 400 Bad Request / 409 Conflict
    - PUT → 200 OK / 404 Not Found
    - DELETE → 204 No Content / 404 Not Found
    - GET → 200 OK / 404 Not Found
servers:
  - url: http://localhost:7051
    description: Public (local)
  - url: http://localhost:7052
    description: Internal (local)
...
paths:
...
  /v1/communications/unsubscribe:
    post:
      tags: [Consent]
      summary: Verify unsubscribe token and record consent opt-out
      operationId: unsubscribe
      x-visibility: public
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnsubscribeRequest"
      responses:
        "204":
          description: Unsubscribed successfully
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Addresses ────────────────────────────────────────────────────────────
...
  /internal/v1/accounts/{id}/communications/unsubscribe-token:
    parameters:
      - $ref: "#/components/parameters/AccountID"
    post:
      tags: [Internal]
      summary: Mint stateless unsubscribe token
      operationId: mintUnsubscribeToken
      x-visibility: private
      security:
        - ServiceJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel]
              properties:
                channel:
                  type: string
      responses:
        "201":
          description: Token minted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MintUnsubscribeTokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

# ---------------------------------------------------------------------------
# Components
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    UserJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User-scoped RS256 JWT issued by komodo-auth-api
    ServiceJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Service-scoped RS256 JWT; must carry at least one `svc:` claim
  parameters:
    AccountID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Account ID
...
  schemas:
...
    MintUnsubscribeTokenResponse:
      type: object
      properties:
        token:
          type: string
    UnsubscribeRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
    ProblemDetail:
      type: object
      description: RFC 7807 Problem+JSON error body
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Forbidden:
      description: JWT present but insufficient scope
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
security:
  - UserJWT: []